update_test_dir

+++ testdir/src/time.c
@@ -6,7 +6,7 @@
 /*   By: pvong <marvin@42lausanne.ch>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/18 10:42:25 by pvong             #+#    #+#             */
-/*   Updated: 2023/05/01 17:12:36 by pvong            ###   ########.fr       */
+/*   Updated: 2023/05/08 10:19:31 by pvong            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ long convert_to_readable_time(struct timeval time)
 	hour = hms / SEC_PER_HOUR;
 	min = (hms % SEC_PER_HOUR) / SEC_PER_MIN;
 	sec = hms % SEC_PER_MIN;
-	printf("%d:%02d:%02d:%d\n", hour, min, sec, time.tv_usec);
+	printf("%d:%02d:%02d:%d\n", hour, min, sec, (int) time.tv_usec);
 	return (hms);
 }
